Add more tests
Still running the same simple 16-bit counter incrementing program on
the Z80 but now the C++ code on the RP2040 verifies that every Z80
memory access is as expected:
* Read or Write access.
* Address of access.
* M1' state of access.

It also checks the value of the 16-bit counter right after the write to
it from the Z80 has completed.

These new tests let me see exactly what causes things to go awry when I
attempt to run the Z80 at higher clock rates. I can now run it
successfully at 700kHz but it fails at 800kHz. These tests helped me
track down the point of failure and see that the fall time of the shift
register was too slow for the 12.5 MHz shift register clocking rate
used when running the Z80 at 800kHz.