Remove redundant shift register clock pulse
The way the PIO code was written before, a full 8 clock cycles were
sent to the pair of shift registers but only 7 are required since the
first bit is ready at the time of latching. It did require adding 2
more PIO instructions to generate these fewer clock pulses though.
When running the Z80 at 750kHz this gives an additional 112ns between
the time the shifts complete and the next Z80 clock falling edge, the
time the CPU has to respond to the request before a wait cycle is
inserted by the Z80.