Use asymmetric shift register clock
The shift register has a maximum clock frequency of 50MHz when run at
5.0V so it can go faster than the 12.5MHz limit that I hit because of
the slow rise/fall times on its outputs. It is only the high portion of
the clock that accounts for this rise/fall time before it is sampled by
the PIO on the falling edge. The low portion of the clock can be much
shorter and not cause any problem. That means if I run the clock
asymmetrically then I can get the benefits of the higher shift rate and
still have time for the outputs to rise/fall. I now run the
AddressStateMachine at a fixed 125MHz which results in a shift register
clock rate of 15.625MHz. The shift clock is based on 8 of the state
machine clock cycles: 6 cycles high and 2 cycles low.

Now when I run the Z80 at 500kHz the code runs fast enough that there
are no wait states needed by the Z80. Higher clock rates in the MHz
work but they just need more wait states for the CPU to catch up.