Handle changing RFSH' pin assignment
Previously the RFSH' pin assignment to GPIO 19 was hardcoded in
the z80_bus.pio assembly language file. rfshPin is provided as a
parameter to Z80_Bus::initAddressStateMachine() and this commit
modifies it to overwrite the machine code for this instruction with
the correct "wait 1 gpio rfshPin side 0b10". This makes it easier to
change in the future like the rest of the pins.