Added debugger support
GDB can now attach over the USB Serial port to debug the Z80. It
supports:
* Memory read/writes
* Register reads/writes
* CTRL+C halting
* Single stepping
* 4 Hardware Breakpoints
* 2 Hardware Watchpoints

I pulled MRI in as a submodule. I had to update it so that it would
support 16-bit registers and addresses like used by the Z80 port of
GDB.

I pulled the Z80 bus handling code out into its own module.

I now run the Z80 memory request handling CPU code on the second RP2040
core so that it doesn't get glitches when the CPU is interrupted by
things like the USB stack.

The old test code is still in main.cpp along with the new debugger
functionality. The g_runMode global can be used to re-enable one of the
tests if desired.