Don't modify stack when halting Z80 into debugger
Previously the halting process would use the Z80's actual stack to store
the registers for grabbing the Z80 context. Now it is redirected to
g_fakeStack instead. This means that the Z80's RAM isn't polluted with
these register values anymore. This might make it easier to debug stack
corruptions or situations where a small stack is used and pushing these
register values on the stack might overwrite application data or code.